+++ public/Data/Dino/carData.json
@@ -1,6 +1,6 @@
 {
   "owner": "박슬기",
-  "number": "116거 1700",
+  "number": "116거1700",
   "car_name": "캐스퍼",
   "trim": "1.0 터보 모던",
   "model_year": 2023,
