+++ src/Pages/Estimate/Estimate.js
@@ -1,31 +1,28 @@
 import React, { useEffect } from 'react';
-import { ProgressBar, Button } from 'react-bootstrap';
+import { ProgressBar } from 'react-bootstrap';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 import {
   currentEstimateState,
   lastEstimateState,
   EstimateCarInfo,
-  UserInputMileageState,
   EstimateCarOption,
 } from '../../atoms';
 import { MdOutlineNavigateNext } from 'react-icons/md';
 import styled, { css } from 'styled-components';
-import Graph from '../Graph/Graph';
+import StateZero from './States/StateZero';
 import StateOne from './States/StateOne';
-import StateFour from './States/StateFour';
-import PhotoCard from '../Sellcar/PhotoCard';
+import StateTwo from './States/StateTwo';
+import StateThree from './States/StateThree';
+import StateFour from './States/StateThree';
+import StateFive from './States/StateFive';
 
 const Estimate = () => {
   const [currentEstimate, setCurrentEstimate] =
     useRecoilState(currentEstimateState);
-  const [estimateCarInfo, setEstimateCarInfo] = useRecoilState(EstimateCarInfo);
+  const setEstimateCarInfo = useSetRecoilState(EstimateCarInfo);
   const setEstimateCarOption = useSetRecoilState(EstimateCarOption);
-  const [userInputMileage, setUserInputMileage] = useRecoilState(
-    UserInputMileageState
-  );
   const [lastEstimate, setLastEstimate] = useRecoilState(lastEstimateState);
-  const { owner, car_name } = estimateCarInfo;
 
   useEffect(() => {
     fetch('http://localhost:3000/Data/Dino/carData.json')
@@ -42,10 +39,6 @@ const Estimate = () => {
         setEstimateCarOption(data);
       });
   }, []);
-
-  const getUserInputMileage = e => {
-    setUserInputMileage(e.target.value);
-  };
 
   const nextProcess = () => {
     setCurrentEstimate(prev => prev + 1);
@@ -82,127 +75,26 @@ const Estimate = () => {
             />
           </ProcessBox>
           {/* STATE 0 : 차량정보 확인 */}
-          <StateOne nextProcess={nextProcess} prevProcess={prevProcess} />
+          {currentEstimate === 0 && <StateZero nextProcess={nextProcess} />}
           {/* STATE 1 : 예상시세 표출 */}
           {currentEstimate === 1 && (
-            <ContentBox>
-              <ContentTitle>
-                <OwnerTag>{owner}</OwnerTag>님의 <CarTag>{car_name}</CarTag> 🚙
-                <br />
-                예상시세는 다음과 같습니다.
-              </ContentTitle>
-              <Graph />
-              <ButtonSet>
-                <PrevButton onClick={prevProcess} variant="primary">
-                  이전
-                </PrevButton>
-                <NextButton onClick={nextProcess} variant="primary">
-                  다음
-                </NextButton>
-              </ButtonSet>
-            </ContentBox>
+            <StateOne prevProcess={prevProcess} nextProcess={nextProcess} />
           )}
           {/* STATE 2 : 주행거리 입력 */}
           {currentEstimate === 2 && (
-            <ContentBox>
-              <ContentTitle>
-                보다 정확한 견적을 위해
-                <br /> 주행거리를 입력해주세요
-              </ContentTitle>
-              <InputBox
-                placeholder="12,345"
-                onChange={e => getUserInputMileage(e)}
-                value={userInputMileage}
-                type="number"
-              />
-              <ButtonSet>
-                <PrevButton onClick={prevProcess} variant="primary">
-                  이전
-                </PrevButton>
-                <NextButton onClick={nextProcess} variant="primary">
-                  다음
-                </NextButton>
-              </ButtonSet>
-            </ContentBox>
+            <StateTwo prevProcess={prevProcess} nextProcess={nextProcess} />
           )}
           {/* STATE 3 : 추가옵션 입력 */}
-          <StateFour nextProcess={nextProcess} prevProcess={prevProcess} />
+          {currentEstimate === 3 && (
+            <StateThree prevProcess={prevProcess} nextProcess={nextProcess} />
+          )}
           {/* STATE 4 : 추가정보 입력 */}
           {currentEstimate === 4 && (
-            <ContentBox>
-              <ContentTitle>
-                보험 외 사고 처리를
-                <br /> 하신 적이 있다면 알려주세요
-              </ContentTitle>
-              <InputBox
-                placeholder="추가입력 사항"
-                // onChange={e => getUserInputOwner(e)}
-                // value={userInputOwner}
-              />
-              <button>보험 외 사고 처리를 한 적이 없어요</button>
-              <ButtonSet>
-                <PrevButton onClick={prevProcess} variant="primary">
-                  이전
-                </PrevButton>
-                <NextButton onClick={nextProcess} variant="primary">
-                  다음
-                </NextButton>
-              </ButtonSet>
-            </ContentBox>
+            <StateFour prevProcess={prevProcess} nextProcess={nextProcess} />
           )}
           {/* STATE 5 : 사진등록 */}
           {currentEstimate === 5 && (
-            <ContentBox>
-              <ContentTitle>차량 사진을 올려주세요</ContentTitle>
-              <PhotoInputContainer>
-                <ContentSubTitle>필수 사진</ContentSubTitle>
-                <ContentSubInfo>
-                  정면, 후면, 측면, 계기판 사진을 올려주세요
-                </ContentSubInfo>
-                <PhotoInputWrapper>
-                  <PhotoInputLine>
-                    {['정면', '후면', '측면', '계기판'].map((value, index) => (
-                      <PhotoCard
-                        key={index}
-                        index={index}
-                        value={value}
-                        // setCarImages={setCarImages}
-                        // carImages={carImages}
-                        // setThumbnails={setThumbnails}
-                        // thumbnails={thumbnails}
-                      />
-                    ))}
-                  </PhotoInputLine>
-                </PhotoInputWrapper>
-                <ContentSubTitle>참고 사진</ContentSubTitle>
-                <ContentSubInfo>
-                  옵션이나 사고부위 등 참고가 될 만한 사진을 올려주세요
-                </ContentSubInfo>
-                <PhotoInputWrapper>
-                  <PhotoInputLine>
-                    {['+', '+', '+', '+'].map((value, index) => (
-                      <PhotoCard
-                        key={index}
-                        index={index}
-                        value={value}
-                        // setCarImages={setCarImages}
-                        // carImages={carImages}
-                        // setThumbnails={setThumbnails}
-                        // thumbnails={thumbnails}
-                      />
-                    ))}
-                  </PhotoInputLine>
-                </PhotoInputWrapper>
-              </PhotoInputContainer>
-              <ButtonSet>
-                <PrevButton onClick={prevProcess} variant="primary">
-                  이전
-                </PrevButton>
-                <NextButton onClick={nextProcess} variant="primary">
-                  다음
-                </NextButton>
-              </ButtonSet>
-            </ContentBox>
+            <StateFive prevProcess={prevProcess} nextProcess={nextProcess} />
           )}
         </EstimateWrapper>
       </BodyWrapper>
@@ -211,81 +103,6 @@ const Estimate = () => {
 };
 
 export default Estimate;
-
-const ContentSubTitle = styled.h3`
-  font-size: medium;
-  font-weight: 600;
-  color: ${({ theme }) => theme.colors.darkGray};
-`;
-
-const ContentSubInfo = styled.div`
-  font-size: small;
-  margin-top: 0.3rem;
-  color: ${({ theme }) => theme.colors.darkGray};
-`;
-
-const PhotoInputContainer = styled.div`
-  /* border: 1px solid black; */
-`;
-
-const PhotoInputWrapper = styled.div`
-  margin: 1rem 0;
-  /* border: 1px solid black; */
-`;
-
-const PhotoInputLine = styled.div`
-  ${({ theme }) => theme.flex.flexBox}
-`;
-
-const OwnerTag = styled.span`
-  color: ${({ theme }) => theme.colors.primaryBlue};
-`;
-
-const CarTag = styled.span`
-  background-color: ${({ theme }) => theme.colors.primaryBlue};
-  color: ${({ theme }) => theme.colors.white};
-  border-radius: 0.2rem;
-  padding: 0 0.3rem;
-  font-size: 22px;
-`;
-
-const InputBox = styled.input`
-  width: 100%;
-  height: 3rem;
-  border: 1px solid ${({ theme }) => theme.colors.disabled};
-  border-radius: 5px;
-  padding: 1em;
-`;
-
-const InputButton = styled(Button)`
-  width: 100%;
-  height: 3rem;
-  border-radius: 100rem;
-  margin-top: 5rem;
-  font-weight: 600;
-`;
-
-const ButtonSet = styled.div`
-  ${({ theme }) => theme.flex.flexBox('', '', 'space-between')}
-  margin-top: 4rem;
-`;
-
-const NextButton = styled(Button)`
-  width: 49%;
-  height: 3rem;
-  border-radius: 100rem;
-  font-weight: 600;
-`;
-
-const PrevButton = styled(Button)`
-  background-color: white;
-  border: 1px solid ${({ theme }) => theme.colors.primaryBlue};
-  color: ${({ theme }) => theme.colors.primaryBlue};
-  width: 49%;
-  height: 3rem;
-  border-radius: 100rem;
-  font-weight: 600;
-`;
 
 const ProcessBox = styled.section`
   ${({ theme }) => theme.flex.flexBox('column')}
@@ -335,21 +152,6 @@ const PercentageBar = styled(ProgressBar)`
   bottom: 0;
 `;
 
-const ContentBox = styled.section`
-  width: 100%;
-  padding: 10%;
-  background-color: white;
-  position: absolute;
-`;
-
-const ContentTitle = styled.h2`
-  font-size: x-large;
-  font-weight: 600;
-  line-height: 1.9rem;
-  margin-bottom: 2.3rem;
-  color: ${({ theme }) => theme.colors.blackC};
-`;
-
 const EstimateWrapper = styled.div`
   width: 100%;
   position: absolute;
@@ -376,10 +178,6 @@ const Background = styled.div`
 `;
 
 const PROCESS_STATE = [
-  // {
-  //   id: 1,
-  //   text: '소유자명',
-  // },
   {
     id: 1,
     text: '차량정보',
